requirements.txt
tensorflow>=2.10.0
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
opencv-python>=4.7.0
pillow>=9.0.0
